requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.2
APScheduler==3.6.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
auto-py-to-exe==2.11.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.9.3
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
cycler==0.11.0
discord.py==1.7.3
distlib==0.3.3
Django==3.2.9
Eel==0.12.4
filelock==3.4.0
Flask==2.0.2
fonttools==4.29.1
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
googlesearch-python==1.0.1
greenlet==1.1.2
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
multidict==5.2.0
numpy==1.21.4
packaging==21.3
pandas==1.3.5
pefile==2021.9.3
Pillow==9.0.1
pipenv==2021.11.23
platformdirs==2.4.0
pycparser==2.21
pygame==2.1.2
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
pyparsing==3.0.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PySide6==6.2.2.1
PySimpleGUI==4.55.1
pyTelegramBotAPI==4.2.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
requests==2.25.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
shiboken6==6.2.2.1
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
sqlparse==0.4.2
telegram==0.0.1
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
Werkzeug==2.0.2
whichcraft==0.6.1
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
